feat(health): Add [fastapi_health](https://github.com/Kludex/fastapi-health) to project and refact code with router
