+++ pyproject.toml
@@ -12,6 +12,7 @@ python-environ = "^0.4.54"
 uvicorn = "^0.14.0"
 fastapi-pagination = "^0.8.3"
 sqlmodel = "^0.0.4"
+fastapi-health = "^0.4.0"
 
 [tool.poetry.dev-dependencies]
 black = "^21.7b0"
