+++ requirements.txt
@@ -3,6 +3,7 @@ chromedriver-binary==92.0.4515.107.0; python_version >= "3.8" and python_version
 click==8.0.1; python_version >= "3.6"
 colorama==0.4.4; python_version >= "3.6" and python_full_version < "3.0.0" and platform_system == "Windows" or platform_system == "Windows" and python_version >= "3.6" and python_full_version >= "3.5.0"
 ecoindex==1.6.0; python_version >= "3.8" and python_version < "4.0"
+fastapi-health==0.4.0; python_full_version >= "3.6.1" and python_full_version < "4.0.0"
 fastapi-pagination==0.8.3; python_version >= "3.7" and python_version < "4.0"
 fastapi==0.68.1; python_version >= "3.6"
 greenlet==1.1.1; python_version >= "3.8" and platform_machine in "x86_64 X86_64 aarch64 AARCH64 ppc64le PPC64LE amd64 AMD64 win32 WIN32" and python_full_version >= "3.6.1" and python_full_version < "4.0.0" and python_version < "4.0"
